fix(dashboard): share leaderboard usd conversion
